+++ internal/cli/package_darwin.go
@@ -7,4 +7,5 @@ import "github.com/yasyf/cc-pool/internal/statuspackage"
 var (
 	installPackage   = statuspackage.Install
 	uninstallPackage = statuspackage.Uninstall
+	resetPackage     = statuspackage.Reset
 )